dhclient.c r1.569 started writing some files before calling daemon(), which involved doing a write_option_db() call at a different point in bind_lease().
However one of the parameters was in a variable that was also used for
iterating a TAILQ so this did not have the intended contents (it was
always NULL at the end of TAILQ_FOREACH*, resulting in a null deref
when dhclient -L is used).

Fix by using a new temporary variable.  ok krw@